Add scripts/plugin_loader.js, which runs after bootstrap.
This code does an XHR for cordova_plugins.json. If it finds the file, it
will load the plugins specified in it. If the response is empty or the
XHR fails, it does nothing.

This creates a new onPluginsReady event that fires either when the
plugins are all loaded or there are no plugins to load. The bootstrap is
waiting on this now, as well as onDOMContentLoaded and onNativeReady.